requirements.txt
streamlit==1.54.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.14
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.1
faiss-cpu==1.13.2
huggingface_hub==0.36.2
sentence-transformers==5.2.3
python-dotenv==1.2.1
pypdf==6.7.2
